requirements.txt
Babel==2.3.4
blinker==1.4
click==6.6
decorator==4.0.10
et-xmlfile==1.0.1
Flask==0.11.1
Flask-OAuthlib==0.9.3
Flask-SQLAlchemy==2.1
Flask-SSLify==0.1.5
Flask-Table==0.2.10
Flask-Uploads==0.2.0
Flask-WTF==0.12
ipython-genutils==0.1.0
itsdangerous==0.24
jdcal==1.2
Jinja2==2.8
MarkupSafe==0.23
oauthlib==1.1.2
openpyxl==2.3.5
pbr==1.10.0
pexpect==4.1.0
pickleshare==0.7.2
pipdeptree==0.6.0
ptyprocess==0.5.1
pytz==2016.4
requests==2.10.0
requests-oauthlib==0.6.1
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.0.13
sqlalchemy-migrate==0.10.0
sqlparse==0.1.19
Tempita==0.5.2
traitlets==4.2.1
Werkzeug==0.11.10
WTForms==2.1
